requirements.txt
# =============================================================================
# LinkedIn Engagement Tracker - Dependências Python
# =============================================================================
# Instale com:
#   pip install -r requirements.txt
#   playwright install chromium
# =============================================================================

# Browser automation
playwright>=1.40.0

# Database ORM
sqlalchemy>=2.0.0

# Data manipulation
pandas>=2.0.0

# Dashboard
streamlit>=1.30.0
plotly>=5.18.0

# Environment variables
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Type checking (optional, for development)
mypy>=1.7.0
